Move flight test. Add assertion for enabled apis
